+++ NEWS.md
@@ -2,6 +2,7 @@ Conserve 0.1.1 not released yet
 ===============================
 
 * Start a manpage.
+* Rename `init-archive` to just `init`.
 
 Conserve 0.1.0 2013-10-01
 =========================